for a vector of indices into a SummarizedExperiment, extract ICCs for a GLS model

for a vector of indices into a SummarizedExperiment, extract ICCs for a GLS model

get_iccs(
  se,
  inds = 1:10,
  makedf.in = makedf_copdgene,
  fixedfmla = beta ~ race + gender + smkstat,
  ...
)

Arguments

se
SummarizedExperiment

inds
numeric() row indices

makedf.in
function that generates a data.frame for use by nlme:gls, consumes se and dots, must return a data.frame with beta and id as columns, fixedfmla will use beta and any columns other than id to build model

fixedfmla
formula with beta as dependent variable

...
passed to makedf.in, within sapply

Examples

requireNamespace("HDF5Array")

#> Loading required namespace: HDF5Array

requireNamespace("SummarizedExperiment")
se6 = HDF5Array::loadHDF5SummarizedExperiment(system.file("chr6_hdf5", package="meffilTools"))
makedf_simple = function(se, ...) {
  stopifnot(nrow(se)==1)
  data.frame(beta=as.numeric(SummarizedExperiment::assay(se)),
       id=as.numeric(factor(se$Sample_Group)))
}
gg = get_iccs(se6, inds=1:5, makedf.in=makedf_simple, fixedfmla=beta~1)
gg

#> cg14092853 cg03238340 cg00948906 cg25744450 cg14728017 
#>  0.2624650  0.7858386  0.6377644  0.2421822 -0.2500000